Move the std module up to root
It is not a bunch of manually constructed `flowrep.prospective.atomic_recipe.AtomicRecipe` objects anymore, but `@flowrep.parsers.atomic_parser.atomic`-decorated functions. I didn't like the `prospective` module cross-talking with the `parser` module like that, so it couldn't stay in `prospective`, and it doesn't actually contribute to parsing so it doesn't really live in `parser`. It's just a sort of useful utility, so moving it up to the root module seems best.
